refactor, working code
