+++ app/lib/__init__.py
@@ -1,2 +1,2 @@
 # __init__.py
-from .ingest_tools import accessioner
+from .dev_tools import accessioner
