+++ app/lib/statusCrawl.py
@@ -1,0 +1,172 @@
+'''
+crawls dir in sources_config to find accessions in system and where they are
+
+The config defines the status (or range of possible statuses) of a given path
+The status_type_order must be a linear definition of status order
+'''
+__author__ = "hewgreen"
+__license__ = "Apache 2.0"
+__date__ = "15/07/2019"
+
+import json
+from datetime import datetime
+import os
+import re
+from collections import defaultdict
+from tqdm import tqdm
+import glob
+
+class atlas_status:
+    def __init__(self, sources_config, status_type_order):
+
+        # configuration
+        with open(sources_config) as f:
+            self.sources_config = json.load(f)
+        self.status_type_order = status_type_order
+        self.accession_regex = re.compile(
+            '^E-(GEOD|MTAB|PROT|ENAD|AFMX|CURD|EHCA|MEXP|TABM|NASC|ERAD|GEUV|JJRD|ATMX|HCAD|MIMR|CBIL|MAXD)-[0-9]+$')
+
+        # status tracking
+        self.status_types = self.get_status_types()
+
+        assert set(self.status_types) == set(
+            self.status_type_order), 'Unrecognised status in config. Please update status type order list.'
+        self.timestamp = datetime.fromtimestamp(datetime.now().timestamp()).isoformat()
+        print('Initialised {}'.format(self.timestamp))
+
+        # accession search
+        accession_search = self.accession_search() # scans dir in config to find '*.idf.txt' or accession directories
+        self.all_primary_accessions = accession_search[0]
+        self.found_accessions = accession_search[1]
+        self.accession_final_status = self.status_tracker() # determines status of each dataset based on location of files
+
+        get_min_max_status = self.get_min_max_status() # sets two variables with min and max status based on status_type_order
+        self.accession_min_status = get_min_max_status[0]
+        self.accession_max_status = get_min_max_status[1]
+
+        latest_idf_sdrf = self.get_latest_idf_sdrf()  # finds path to latest idf and sdrf file
+        self.idf_path_by_accession = latest_idf_sdrf[0]
+        self.sdrf_path_by_accession = latest_idf_sdrf[1]
+        self.path_by_accession = latest_idf_sdrf[2]
+
+    def get_status_types(self):
+        status_types = set()
+        for x, y in self.sources_config.items():
+            for n in y['stage']:
+                status_types.add(n)
+        return list(status_types)
+
+    def accession_search(self):
+        print('Performing accession search {}'.format(
+            datetime.fromtimestamp(datetime.now().timestamp()).isoformat()))
+        all_primary_accessions = set()
+        found_accessions = {}
+        counter = 0
+        for path, info in self.sources_config.items():
+            counter += 1
+            print('Searching path {} {}/{}'.format(path, counter, len(self.sources_config)))
+
+            pre_accessions = os.listdir(path)
+            for pre_accession in pre_accessions:
+                if not pre_accession.endswith('.merged.idf.txt'):
+                    accession = pre_accession.strip('.idf.txt')
+                    if self.accession_regex.match(accession):
+                        all_primary_accessions.add(accession)
+                        found_accessions[(path, accession)] = dict(
+                                                                        # config_path=path,
+                                                                        accession=accession,
+                                                                        tech=info.get('tech', None),
+                                                                        stage=info.get('stage', None),
+                                                                        resource=info.get('resource', None),
+                                                                        source=info.get('source', None)
+                        )
+        print('Found {} accessions in {} directories'.format(len(found_accessions), len(self.sources_config)))
+        return all_primary_accessions, found_accessions
+
+    def status_tracker(self):
+        print('Calculating status of each project {}'.format(datetime.fromtimestamp(datetime.now().timestamp()).isoformat()))
+        accession_status = {}
+        accession_status_counter = {}
+        for key, value in self.found_accessions.items():
+            accession = value['accession']
+            stage = value['stage']
+
+            index = self.status_type_order.index(stage[-1])
+            if accession not in accession_status_counter:
+                accession_status_counter[accession] = index
+                accession_status[accession] = ' '.join(stage)
+            elif index > accession_status_counter.get(accession):
+                accession_status_counter[accession] = index
+                accession_status[accession] = ' '.join(stage)
+
+        return accession_status
+
+    def get_min_max_status(self):
+        # Some paths define multiple statuses. This narrows it to the latter most status according to status_type_order
+        accession_min_status = {}
+        accession_max_status = {}
+        for accession, status in self.accession_final_status.items():
+            accession_min_status[accession] = self.status_type_order[min([self.status_type_order.index(x) for x in status.split(' ')])]
+            accession_max_status[accession] = self.status_type_order[max([self.status_type_order.index(x) for x in status.split(' ')])]
+
+        return accession_min_status, accession_max_status
+
+    def get_latest_idf_sdrf(self):
+        print('Getting latest IDF and SDRF paths {}'.format(
+            datetime.fromtimestamp(datetime.now().timestamp()).isoformat()))
+
+        def list_converter(file_list):
+            files_found = {}
+            for filepath in file_list:
+                accession = filepath.split('/')[-1].replace('.idf.txt', '').replace('.sdrf.txt', '')
+                if accession in files_found:
+                    files_found[accession].append(filepath)
+                else:
+                    files_found[accession] = [filepath]
+            return files_found
+
+        def get_ranked_paths():
+            # ranks path by status type, used to estimate out where latest metadata is.
+            preordered_paths = []
+            latest_stage_ranks = []
+            for path, metadata in self.sources_config.items():
+                preordered_paths.append(path)
+                latest_stage_ranks.append(max([self.status_type_order.index(n) for n in metadata.get('stage')]))
+            return [x for _,x in sorted(zip(latest_stage_ranks,preordered_paths))]
+
+        def get_latter_ranked_path(paths_by_accession, ranked_paths):
+            ranked_paths_by_accession = {}
+            for accession, path_list in paths_by_accession.items():
+                if len(path_list) == 1:
+                    latest_path = path_list[0]
+                else:
+                    try:
+                        trunc_paths = [v.split('/E-')[0] for v in path_list]  # remove accession specific endings
+                        path_ranks = [ranked_paths.index(n) for n in trunc_paths]
+                        latest_path = path_list[path_ranks.index(max(path_ranks))]  # get idf/sdrf from latter loc
+                    except ValueError:
+                        continue #todo this needs more investigation in prod got ValueError: '/nfs/production3/ma/home/atlas3-production/singlecell/experiment/ng' is not in list
+                ranked_paths_by_accession[accession] = latest_path
+            return ranked_paths_by_accession
+
+        idf_list_ = []
+        sdrf_list_ = []
+        for path, metadata in tqdm(self.sources_config.items(), unit='Paths in config'):
+            print('IDF/SDRF path finder exploring {}'.format(path))
+            idf_list_ += glob.glob(path + '/*/*.idf.txt') + glob.glob(path + '/*.idf.txt')
+            sdrf_list_ += glob.glob(path + '/*/*.sdrf.txt') + glob.glob(path + '/*.sdrf.txt')
+
+        idf_list = [x for x in idf_list_ if self.accession_regex.match(x.split('/')[-1].replace('.idf.txt', '').replace('.sdrf.txt', ''))]
+        sdrf_list = [x for x in sdrf_list_ if self.accession_regex.match(x.split('/')[-1].replace('.idf.txt', '').replace('.sdrf.txt', ''))]
+
+        ranked_paths = get_ranked_paths()
+        idf_path_by_accession = get_latter_ranked_path(list_converter(idf_list), ranked_paths)
+        sdrf_path_by_accession = get_latter_ranked_path(list_converter(sdrf_list), ranked_paths)
+
+        # get the path where the accession was initially found at
+        paths_by_accession = defaultdict(list)
+        for k, v in self.found_accessions.items():
+            paths_by_accession[v.get('accession')].append(k[0])
+        path_by_accession = get_latter_ranked_path(paths_by_accession, ranked_paths)
+
+        return idf_path_by_accession, sdrf_path_by_accession, path_by_accession
